requirements.txt
black==22.8.0
asgiref==3.5.0
pytz==2022.1
Django==4.1.2
django-cors-headers==3.13.0
django-taggit==3.0.0
djangorestframework==3.14.0
stripe==4.2.0
psycopg2==2.9.3
sqlparse==0.4.2